Move `<Values />` from XYChart/ to be under components/
Since it'll be used by waterfall chart
