+++ frontend/src/metabase/static-viz/components/Values/index.ts
